bots shouldnt get caught in the link filter

+++ src/filters/LinkFilter.js
@@ -6,6 +6,11 @@ class LinkFilter {
     }
 
     filter() {
+
+        // Skip bots
+        if(this.m.author.bot)
+            return;
+
         if (this.message.includes('https://') || this.message.includes('http://') || this.message.includes('www.')) {
             this.m.delete();
             console.log('Link found deleting');
